continued work on problem "Yes, Yes, It's Nonograms"
Implementation 2 is based on implementation 1, ArrayLists were replaced
by arrays where applicable, remaining ArrayLists were trimmed if
possible, values for pixels were changed from byte to boolean type.
None of this fixed the problem of high memory consumption.
After finally doing some math (i feel stupid for not trying this before)
it turns out, that the shear number of possible solutions (in a nearly
worst case) for a single row is way to high for storing in memory
Another approach for generation of possible solutions is required,
without the option for storage.